+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 **Features** 
 * Added: _json-patch_ support, see `JsonPatch`, `JsonPointer`
-* Added: node level bulk modification API, see `JsonNode.Add`, `JsonNode.Remove`
+* Added: node bulk modification API: `JsonNode#patch` + `JsonNode.Insert`, `JsonNode.Remove`
 * Added: test for same information `JsonValue#equivalentTo`
 * Added: test for same definition (ignoring formatting) `JsonValue#identicalTo`
+* Added: `@Validation#acceptNull` to declare that `null` value satisfies required
 
 **Breaking Changes**